Working on junit test for descriptions list in editable lists
