+++ openxavatest/src/main/java/org/openxava/test/model/Appointment2.java
@@ -3,7 +3,6 @@ package org.openxava.test.model;
 import java.util.*;
 
 import javax.persistence.*;
-import javax.validation.constraints.*;
 
 import org.openxava.annotations.*;
 import org.openxava.model.*;
@@ -15,9 +14,8 @@ import org.openxava.model.*;
 
 @Entity
 @Table(name="APPOINTMENT")
-@Tab(properties="time, dateTime, description, peopleCount, type.description, extendedDescription",
-	editableProperties="peopleCount, type.description, extendedDescription") 
-@View(members="time; dateTime; description; peopleCount; type")
+@Tab(properties="time, dateTime, description, amountOfPeople, type.description")
+@View(members="time; dateTime; description; amountOfPeople; type")
 public class Appointment2 extends Identifiable {
 	
 	@Stereotype("DATETIME")
@@ -31,12 +29,7 @@ public class Appointment2 extends Identifiable {
 	@Column(length=60) @Required
 	private String description;
 	
-	@Column(name="AMOUNTOFPEOPLE")
-	@Max(13) 
-	private int peopleCount;
-
-	@org.hibernate.annotations.Formula("DESCRIPTION || ' ' || AMOUNTOFPEOPLE")
-	private String extendedDescription; 
+	private int amountOfPeople;
 	
 	@DescriptionsList
 	@ManyToOne(fetch=FetchType.LAZY)
@@ -66,6 +59,14 @@ public class Appointment2 extends Identifiable {
 		this.description = description;
 	}
 
+	public int getAmountOfPeople() {
+		return amountOfPeople;
+	}
+
+	public void setAmountOfPeople(int amountOfPeople) {
+		this.amountOfPeople = amountOfPeople;
+	}
+
 	public AppointmentType getType() {
 		return type;
 	}
@@ -73,21 +74,4 @@ public class Appointment2 extends Identifiable {
 	public void setType(AppointmentType type) {
 		this.type = type;
 	}
-	
-	public String getExtendedDescription() {
-		return extendedDescription;
-	}
-
-	public void setExtendedDescription(String extendedDescription) {
-		this.extendedDescription = extendedDescription;
-	}
-
-	public int getPeopleCount() {
-		return peopleCount;
-	}
-
-	public void setPeopleCount(int peopleCount) {
-		this.peopleCount = peopleCount;
-	}
-
 }
