+++ openxavatest/src/main/java/org/openxava/test/model/Product2.java
@@ -121,8 +121,8 @@ import org.openxava.test.validators.*;
 })
 @Tab(properties="number, description, family.description, subfamily.description, unitPrice, unitPriceWithTax, subfamily.family.description") 
 @Tab(name="Editable", // tmr Si nos quedamos con esto, quitar lo de Appointment2
-	properties="number, description, family.description, unitPrice, warehouse.name",
-	editableProperties="family.description, unitPrice, warehouse.name"
+	properties="number, description, family.description, unitPrice, warehouse.name, extendedDescription",
+	editableProperties="family.description, unitPrice, warehouse.name, extendedDescription"
 )
 public class Product2 {
 	
@@ -209,6 +209,9 @@ public class Product2 {
 	// Only to show in view
 	@Transient @Stereotype("SUBFAMILY_DEPENDS_REFERENCE")
 	private int subfamilyNumber; 
+	
+	@org.hibernate.annotations.Formula("NUMBER || ' - ' || DESCRIPTION")
+	private String extendedDescription; 
 	
 	@Depends("unitPrice") 
 	@Max(999999999999999999L) 	
@@ -345,6 +348,14 @@ public class Product2 {
 
 	public void setColor(Color color) {
 		this.color = color;
+	}
+
+	public String getExtendedDescription() {
+		return extendedDescription;
+	}
+
+	public void setExtendedDescription(String extendedDescription) {
+		this.extendedDescription = extendedDescription;
 	}
